URI parser changes to support decoding of percentage encoded special characters
According to RFC8040, yang list node keys that contains specical
chararcters such as "/" or ":" are percentage encoded by the restconf
client when sending URIs to the restconf server.

I have made the changes accordingly on restconf server to pass
pass percentage encoded key strings to yang runtime. This submission
is to change the URI utility code to accordingly. With this submission,
restconf GET/POST operations will be able to support list nodes whose
keys contain "/" or ":".

Note that this submission does not depend on the restconf server
changes, and thus can be merged in directly.

Change-Id: I000e31ada9c4e0ad1f0196c45487b326e5a7741c